+++ www/apps/api-reference/markdown/store.mdx
@@ -5,7 +5,8 @@ import {
   DividedMarkdownCode
 } from "@/layouts/DividedMarkdown/Sections"
 import Section from "@/components/Section"
-import { Feedback, CodeTabs, CodeTab, H1 } from "docs-ui"
+import { CodeTabs, CodeTab, H1 } from "docs-ui"
+import { Feedback } from "@/components/Feedback"
 
 import ClientLibraries from "./client-libraries.mdx"
 
@@ -24,14 +25,10 @@ This API reference includes Medusa v2's Store APIs, which are REST APIs exposed 
 All API Routes are prefixed with `/store`. So, during development, the API Routes will be available under the path `http://localhost:9000/store`. For production, replace `http://localhost:9000` with your Medusa application URL.
 
 <Feedback
-  event="survey_api-ref"
   extraData={{
-    area: "store",
     section: "introduction"
   }}
-  pathName="/api/store"
   question="Was this section helpful?"
-  vertical={true}
 />
 
 </DividedMarkdownContent>
@@ -264,14 +261,10 @@ fetch(`<BACKEND_URL>/store/products`, {
 <DividedMarkdownLayout addYSpacing>
 
 <Feedback
-  event="survey_api-ref"
   extraData={{
-    area: "store",
     section: "authentication-cookie"
   }}
-  pathName="/api/store"
   question="Was this section helpful?"
-  vertical={true}
 />
 
 </DividedMarkdownLayout>
@@ -349,14 +342,10 @@ Where `{your_publishable_api_key}` is the token of the publishable API key. When
 <DividedMarkdownLayout>
 
 <Feedback
-  event="survey_api-ref"
   extraData={{
-    area: "store",
     section: "publishable-api-key"
   }}
-  pathName="/api/store"
   question="Was this section helpful?"
-  vertical={true}
 />
 
 </DividedMarkdownLayout>
@@ -421,14 +410,10 @@ x-no-compression: false
 <DividedMarkdownLayout>
 
 <Feedback
-  event="survey_api-ref"
   extraData={{
-    area: "store",
     section: "http-compression"
   }}
-  pathName="/api/store"
   question="Was this section helpful?"
-  vertical={true}
 />
 
 </DividedMarkdownLayout>
@@ -624,14 +609,10 @@ sdk.store.cart.updateLineItem(
 To remove a property from the `metadata`, pass the property in the request body with an empty string value. This will remove the property from the `metadata` without affecting other properties.
 
 <Feedback
-  event="survey_api-ref"
   extraData={{
-    area: "store",
     section: "metadata"
   }}
-  pathName="/api/store"
   question="Was this section helpful?"
-  vertical={true}
 />
 
 </DividedMarkdownContent>
@@ -925,14 +906,10 @@ However, some API routes restrict the fields and relations you can retrieve. To 
 <DividedMarkdownLayout>
 
 <Feedback
-  event="survey_api-ref"
   extraData={{
-    area: "store",
     section: "select-fields"
   }}
-  pathName="/api/store"
   question="Was this section helpful?"
-  vertical={true}
 />
 
 </DividedMarkdownLayout>
@@ -1289,14 +1266,10 @@ curl -g "http://localhost:9000/store/products?created_at[$lt]=2023-02-17&created
 <DividedMarkdownLayout>
 
 <Feedback
-  event="survey_api-ref"
   extraData={{
-    area: "store",
     section: "query-parameters"
   }}
-  pathName="/api/store"
   question="Was this section helpful?"
-  vertical={true}
 />
 
 </DividedMarkdownLayout>
@@ -1468,14 +1441,10 @@ This sorts the products by their `created_at` field in the descending order.
 <DividedMarkdownLayout>
 
 <Feedback
-  event="survey_api-ref"
   extraData={{
-    area: "store",
     section: "pagination"
   }}
-  pathName="/api/store"
   question="Was this section helpful?"
-  vertical={true}
 />
 
 </DividedMarkdownLayout>
@@ -1499,14 +1468,10 @@ This is useful if you want to extend an API route and pass additional data or pe
 Refer to [this guide](!docs!/learn/customization/extend-features/extend-create-product) to find an example of extending an API route.
 
 <Feedback
-  event="survey_api-ref"
   extraData={{
-    area: "store",
     section: "workflows"
   }}
-  pathName="/api/store"
   question="Was this section helpful?"
-  vertical={true}
 />
 
 </DividedMarkdownContent>
