+++ www/apps/api-reference/package.json
@@ -34,6 +34,7 @@
     "openapi-sampler": "^1.3.1",
     "pluralize": "^8.0.0",
     "postcss": "8.4.27",
+    "posthog-js": "^1.269.1",
     "prism-react-renderer": "2.4.0",
     "react": "rc",
     "react-dom": "rc",
