+++ www/apps/book/instrumentation-client.ts
@@ -1,0 +1,7 @@
+import posthog from "posthog-js"
+
+posthog.init(process.env.NEXT_PUBLIC_POSTHOG_KEY!, {
+  api_host: process.env.NEXT_PUBLIC_POSTHOG_HOST,
+  person_profiles: "always",
+  defaults: "2025-05-24",
+})
