+++ www/apps/book/package.json
@@ -23,6 +23,7 @@
     "docs-ui": "*",
     "docs-utils": "*",
     "next": "15.3.5",
+    "posthog-js": "^1.269.1",
     "react": "rc",
     "react-dom": "rc",
     "rehype-mdx-code-props": "^2.0.0",
