+++ www/apps/book/providers/index.tsx
@@ -26,8 +26,6 @@ const Providers = ({ children, aiAssistantProps = {} }: ProvidersProps) => {
     <AnalyticsProvider
       segmentWriteKey={process.env.NEXT_PUBLIC_SEGMENT_API_KEY}
       reoDevKey={process.env.NEXT_PUBLIC_REO_DEV_CLIENT_ID}
-      postHogKey={process.env.NEXT_PUBLIC_POSTHOG_KEY}
-      postHogApiHost={process.env.NEXT_PUBLIC_POSTHOG_HOST}
     >
       <SiteConfigProvider config={config}>
         <LearningPathProvider>
